chore: move config bindings to separate wrapper
